requirements.txt
asn1crypto==0.24.0
astroid==1.6.0
backports.functools-lru-cache==1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
configparser==3.5.0
cryptography==2.1.4
enum34==1.1.6
enum-compat==0.0.2
gTTS==1.2.2
gTTS-token==1.1.1
idna==2.6
ipaddress==1.0.19
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
netifaces==0.10.6
protobuf==3.5.1
PyChromecast==1.0.3
pycparser==2.18
PyJWT==1.5.3
pylint==1.8.1
pyOpenSSL==17.5.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
RPi.GPIO==0.6.3
singledispatch==3.4.0.3
six==1.11.0
twilio==6.10.0
urllib3==1.22
wrapt==1.10.11
zeroconf==0.19.1
